fixed some issues caused by merge

+++ server/helping-hands-backend/src/main/java/com/helpinghands/backendPrototype/Controllers/AuthenticationController.java
@@ -21,7 +21,7 @@ public class AuthenticationController {
     private static final String userSessionKey = "user";
 
     public User getUserFromSession(HttpSession session) {
-        Integer userId = (Integer) session.getAttribute(userSessionKey);
+        Long userId = (Long) session.getAttribute(userSessionKey);
         if (userId == null) {
             return null;
         }
